Moving standardization from crud to chemical_utils
